+++ src/letter/letter.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
-import { LettersController } from './controller/letters.controller';
-import { LettersService } from './service/letters.service';
+import { LettersController } from './controller/letter.controller';
+import { LettersService } from './service/letter.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { LetterRepository } from './repository/letters.repository';
-import { Letter } from './model/letters.entity';
+import { LetterRepository } from './repository/letter.repository';
+import { Letter } from './model/letter.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Letter])],
