+++ src/letter/model/letter-scope.enum.ts
